requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask_bcrypt
flask_wtf
flask_login
mysqlclient
email_validator
python-dotenv
firebase_admin